Harden selective upstream integrations
Complete Studio-specific compatibility coverage for upstream 92bdb6e, efb58c8, and b677cdb with architecture typing, lightweight Mage-Flow runtime tests, sync-safe atomic staging, and dtype-safe bounded UIntX unpacking.

+++ toolkit/util/uintx_quant.py
@@ -115,7 +115,7 @@ def unpack_uintx_range(
             slots = torch.div(indices, nbytes, rounding_mode="floor")
             vals = torch.bitwise_right_shift(
                 packed[pos + byte_indices],
-                slots * k,
+                (slots * k).to(torch.uint8),
             )
             vals.bitwise_and_((1 << k) - 1)
             out |= vals.bitwise_left_shift(offset)
